Clarify comment about babel-preset-env external config support
Previously this comment read:
> no config support, only tool option

However, this was confusing, because if you follow the link to the
babel-preset-env repository, you will find documentation about support
for external config in package.json or browserlist files:
https://github.com/babel/babel-preset-env/blame/927a3b521907fce260898208a3d30c1694917730/README.md#L108-L149

That's because the master branch of the babel-preset-env repository is
for version 2.0, while the README for the latest 1.x release is only
available on the 1.x branch.

Version 2.0 is currently still in beta, but it includes support for
reading an external config in package.json or browserslist files (via
https://github.com/babel/babel-preset-env/pull/161).

With this in mind, I've updated this comment to:
> external config in package.json or browserslist files supported in v2.0+

in an effort to remove this confusion. Once version 2.0 gets out of
beta, we can probably remove this comment, but until then I think it's
helpful, and lets users know that if they want external config support
they can try the latest beta release of babel-present-env v2.0.