Relax validation when extending config from scoped packages
The SCOPED_CONFIG__PATTERN regex and the
'`.` not allowed in Browserslist config name.' rules disallowed
extending a browserslist from a package called
`@example.com/browserslist-config` even though that is a perfectly valid
name for a scoped package.

I loosened up the regex SCOPED_CONFIG__PATTERN so it would match an npm
org named example.com and modifed the other rule to remove a scope
before it checks for dots in the extend reference.