Updating node engines
### Problem
I ran into a particular issue that my node version was 13.6 and it was throwing an error because how is declared in our engines. Updating to the 13.7 seems to fixed the issue.

### Solution
We can update the version to ^13 and include 14 since it's already out.

Thoughts? I have not  test this PR yet.