Add "since COVID" and "post-COVID" queries
The cultural importance of COVID cannot be overstated. It marks the
beginning of a new era, one in which an action as simple as going
outside might be deadly. Because of the major changes it introduced into
our lives, as well as trauma that cannot be forgotten, many use it as a
date signifier even in contexts unrelated to the pandemic itself.